make the cursor a little thicker
I'm testing this by moving the cursor around with my eyes closed, then
starting a stopwatch as I open my eyes. This seems to help a bit. I'm
able to acquire the cursor in 2s. At least the 10s outliers I used to
have with the circle or thin line don't seem to be happening.