new test
For commit ff88a2a927 which fixed a regression caused by commit
e51ce12969.

I'm trying to provide enough guardrails for myself and future readers
without causing a combinatorial explosion in tests. The previous commit
was able to get more value out of existing tests, but this test feels
necessary. And useful in general without reference to a specific bug.