bugfix in search UI
This one is ancient and it affects every single one of my forks,
including the whole lines2 lineage. The corner case: searching for empty
string.

In the process I've also cleaned up edit.check_locs on initialization to
only modify cursor if it can find a legal place for it.

In general I should be more careful about mutating the cursor. Just
adding 1 to it is irresponsible.